Implements range modifier for time tracker
Hooks in the range modifier. This leverages the individual selection
from the form on the timer tracker. When the BE is hooked up, I believe
that we will leverage state and let that dictate the initial range;
followed by specific modifications on that form (if desired)

Also begins to fixup a stakeholder issue:
* Selecting a time range was confusing

8 - 11AM was showing 8 - 12pm (because 11 was "inclusive of the full
hour")

I am slowly improving that, as well as showing a "terminal line" for the
UI on the time tracker to indicate that it is _through_ that time.